yargitay-scraper-api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
loguru
selenium
tenacity
# sqlalchemy # Kaldırıldı
# psycopg2-binary # Kaldırıldı
passlib[bcrypt]
python-dotenv
slowapi
python-multipart
uuid
beanie # Eklendi
motor